requirements.txt
websocket-client==1.8.0
python-dotenv==1.1.1
pyaudio==0.2.14
websockets==15.0.1